Fix test end points
