+++ src/components/NewAppointment.js
@@ -3,7 +3,6 @@ import { useNavigate } from 'react-router-dom';
 import { createAppointment } from '../redux/doctors/doctorSlice';
 import { useAppDispatch, useAppSelector } from '../hooks/hooks';
 
-
 function NewAppointment() {
   const dispatch = useDispatch();
 
@@ -11,15 +10,8 @@ function NewAppointment() {
   const [date, setDate] = useState('');
   const [city, setCity] = useState('');
 
-function NewDoctor() {
-  const user = useAppSelector((state) => state.user.userData.user);
-  const dispatch = useAppDispatch();
-  const navigate = useNavigate();
-
-
   const implementAppointment = (e) => {
     e.preventDefault();
-
 
     const data = {
 
@@ -33,7 +25,6 @@ function NewDoctor() {
     formData.append('appointment[date]', e.target.date.value);
     formData.append('appointment[city]', e.target.city.value);
 
-
     dispatch(createAppointment(JSON.parse(user).id, formData)).then((result) => {
       if (result && result.error) return;
       navigate('/doctors');
@@ -43,22 +34,8 @@ function NewDoctor() {
 
   return (
     <div>
-
-      <h1>Add Doctor</h1>
-      <form className="add-form w-60" onSubmit={addAppointment}>
-        <input
-
       <h1>Add Appointment</h1>
       <form onSubmit={implementAppointment}>
-        {/* <input
-          id="doctor"
-          placeholder="doctor"
-          type="text"
-          name="doctor"
-          className="form-control mb-3"
-          required
-          autoComplete="doctor"
-        /> */}
 
         <input
           id="date"
